+++ src/main/java/eu/dissco/core/digitalspecimenprocessor/web/WebClientUtils.java
@@ -1,4 +1,4 @@
-package eu.dissco.core.digitalspecimenprocessor.component;
+package eu.dissco.core.digitalspecimenprocessor.web;
 
 import org.springframework.web.reactive.function.client.WebClientResponseException;
